Move dynamic_tool_models to tool_util_models
