+++ lib/galaxy/webapps/galaxy/api/dynamic_tools.py
@@ -15,14 +15,14 @@ from galaxy.managers.tools import (
 )
 from galaxy.model import User
 from galaxy.schema.fields import DecodedDatabaseIdField
-from galaxy.tool_util.dynamic_tool_models import (
+from galaxy.tool_util.parameters import input_models_for_tool_source
+from galaxy.tool_util.parameters.convert import cwl_runtime_model
+from galaxy.tool_util.parser.yaml import YamlToolSource
+from galaxy.tool_util_models.dynamic_tool_models import (
     DynamicToolPayload,
     DynamicUnprivilegedToolCreatePayload,
     UnprivilegedToolResponse,
 )
-from galaxy.tool_util.parameters import input_models_for_tool_source
-from galaxy.tool_util.parameters.convert import cwl_runtime_model
-from galaxy.tool_util.parser.yaml import YamlToolSource
 from galaxy.webapps.galaxy.api import (
     depends,
     DependsOnTrans,
