+++ lib/galaxy_test/base/populators.py
@@ -102,7 +102,6 @@ from galaxy.tool_util.cwl.util import (
     output_to_cwl_json,
     tool_response_to_output,
 )
-from galaxy.tool_util.dynamic_tool_models import DynamicUnprivilegedToolCreatePayload
 from galaxy.tool_util.verify.test_data import TestDataResolver
 from galaxy.tool_util.verify.wait import (
     timeout_type,
@@ -110,6 +109,7 @@ from galaxy.tool_util.verify.wait import (
     wait_on as tool_util_wait_on,
 )
 from galaxy.tool_util_models import UserToolSource
+from galaxy.tool_util_models.dynamic_tool_models import DynamicUnprivilegedToolCreatePayload
 from galaxy.util import (
     DEFAULT_SOCKET_TIMEOUT,
     galaxy_root_path,
